Refactor caching into reaction_tracking and implement a cache length cap
