+++ cogs/vote_tracking.py
@@ -1,4 +1,3 @@
-from discord import TextChannel, Message, User
 from discord.ext import commands
 
 from config import config
@@ -15,16 +14,14 @@ class VoteTracking(commands.Cog, name='Vote Tracking'):
 
     def __init__(self, bot) -> None:
         self.bot = bot
-        self.cachedChannels = {}
-        self.cachedMessages = {}
-        self.cachedUsers = {}
+        self.cache = reaction_tracking.MemoizeCache(bot)
 
     @commands.Cog.listener('on_raw_reaction_add')
     async def on_reaction_add(self, event):
-        channel = await self.get_channel(event.channel_id)
-        message = await self.get_message(channel, event.message_id)
+        channel = await self.cache.get_channel(event.channel_id)
+        message = await self.cache.get_message(channel, event.message_id)
         created_at = message.created_at
-        user = await self.get_user(event.user_id)
+        user = await self.cache.get_user(event.user_id)
         username = user.name
 
         emoji = reaction_tracking.format_emoji(event.emoji)
@@ -68,33 +65,6 @@ class VoteTracking(commands.Cog, name='Vote Tracking'):
         result = reaction_tracking.create_channel_tracking_relationship(ctx.channel.id, textChannel.id)
         
         await ctx.send(result)
-
-    async def get_channel(self, id: int) -> TextChannel:
-        if id not in self.cachedChannels:
-            print(f'Caching channel {id}')
-            self.cachedChannels[id] = await self.bot.fetch_channel(id)
-        else:
-            print(f'Getting cached channel {id}')
-
-        return self.cachedChannels[id]
-
-    async def get_message(self, channel: TextChannel, id: int) -> Message:
-        if id not in self.cachedMessages:
-            print(f'Caching message {id}')
-            self.cachedMessages[id] = await channel.fetch_message(id)
-        else:
-            print(f'Getting cached message {id}')
-
-        return self.cachedMessages[id]
-
-    async def get_user(self, id: int) -> User:
-        if id not in self.cachedUsers:
-            print(f'Caching user {id}')
-            self.cachedUsers[id] = await self.bot.fetch_user(id)
-        else:
-            print(f'Getting cached user {id}')
-
-        return self.cachedUsers[id]
 
 
 async def setup(bot):
