+++ config/config.py
@@ -133,3 +133,4 @@ PHASE_GROUPS = [7]
 
 # Reaction Tracking
 REACTION_TRACKING_EXPIRY_DAYS = 7
+MAX_CACHE_LENGTH = 100
